% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_fct.R
\name{check_project}
\alias{check_project}
\title{Title}
\usage{
check_project(list_input, project)
}
\arguments{
\item{list_input}{list dataset list}

\item{project}{project id}
}
\value{
filtered dataset list by project
}
\description{
Title
}
