% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_fct.R
\name{create_download_request_body}
\alias{create_download_request_body}
\title{Title}
\usage{
create_download_request_body(lexis_ds, file_path = "/")
}
\arguments{
\item{file_path}{}
}
\description{
Title
}
