% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_file_from_dataset}
\alias{download_file_from_dataset}
\title{Title}
\usage{
download_file_from_dataset(
  lexis_oauth_token,
  lexis_ds,
  destination_path,
  source_file_path = "/",
  max_retries = 100,
  verbosity = 0,
  extract = FALSE
)
}
\arguments{
\item{extract}{}
}
\description{
Title
}
