% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{extract_dataset}
\alias{extract_dataset}
\title{Title}
\usage{
extract_dataset(lexis_ds_list, search_value, ...)
}
\arguments{
\item{lexis_ds_list}{list of dataset made by get_dataset_list() function}

\item{search_value}{character, what should be matched value}

\item{...}{values, which should be passed to purrr::pluck() function used to extract values from nested lists}
}
\description{
Title
}
