% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_file_list}
\alias{get_dataset_file_list}
\title{Title}
\usage{
get_dataset_file_list(
  lexis_oauth_token,
  internalID,
  project,
  path = "",
  access = "project",
  recursive = TRUE,
  zone = "IT4ILexisZone"
)
}
\arguments{
\item{zone}{}
}
\description{
Title
}
