% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_lexis_oauth_token}
\alias{get_lexis_oauth_token}
\title{Title}
\usage{
get_lexis_oauth_token(
  host_name = "localhost",
  host_ip = "127.0.0.1",
  port = httpuv::randomPort()
)
}
\arguments{
\item{port}{}
}
\description{
Title
}
