% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_lexis_user_info}
\alias{get_lexis_user_info}
\title{Title}
\usage{
get_lexis_user_info(
  lexis_oauth_token,
  base_path = "https://aai.lexis.tech",
  dry_run = FALSE,
  verbosity = 0
)
}
\arguments{
\item{verbosity}{}
}
\description{
Title
}
