% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_request.R
\name{lexis_request_post}
\alias{lexis_request_post}
\title{Title}
\usage{
lexis_request_post(
  lexis_oauth_token,
  request_api_path,
  req_method = NULL,
  req_query = NULL,
  req_body_json = NULL,
  resp_body_type = "",
  base_path = "https://api.lexis.tech",
  dry_run = FALSE,
  verbosity = 0
)
}
\arguments{
\item{verbosity}{}
}
\description{
Title
}
