% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{upload_dataset}
\alias{upload_dataset}
\title{Title}
\usage{
upload_dataset(
  file_path = NULL,
  file_stream = NULL,
  lexis_oauth_token,
  metadata,
  url = "https://api.lexis.tech/api/v0.2/transfer/upload",
  chunk_size = 2048,
  retries = 10,
  retry_delay = 15,
  show_progress = TRUE
)
}
\arguments{
\item{show_progress}{}
}
\description{
Title
}
